Move the histogram class into its own files.
As it says. This is in preparation for bringing this code up-to-date
with the JIT CC.
